Wrote the code for the functioning of an android app that stores data for multiple users and displays the data using 2 storage methods: Firebase and Shared Preferences.

+++ app/src/main/java/com/rishabh/monkhoodassignment/mvvm/ViewModel.kt
@@ -2,6 +2,7 @@ package com.rishabh.monkhoodassignment.mvvm
 
 import android.content.Context
 import android.content.SharedPreferences
+import android.util.Log
 import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -18,8 +19,8 @@ class ViewModel: ViewModel() {
     val userListSharedPreferences = MutableLiveData<List<Users>>()
 
     val firestore = FirebaseFirestore.getInstance()
-    val sharedPreferences: SharedPreferences = MyApplication.getAppContext().getSharedPreferences("Users",
-        Context.MODE_PRIVATE)
+    val sharedPreferences: SharedPreferences = MyApplication.getAppContext()
+        .getSharedPreferences("Users", Context.MODE_PRIVATE)
 
     fun getUsersFromFirebase() : LiveData<List<Users>> {
         viewModelScope.launch(Dispatchers.IO) {
@@ -41,32 +42,44 @@ class ViewModel: ViewModel() {
         return userListFirebase
     }
 
-    fun deleteUserFromFirebase(userId : String){
+    fun deleteUserFromFirebase(userId: String) {
+        if (userId.isNullOrEmpty()) {
+            Toast.makeText(MyApplication.getAppContext(), "Invalid user ID!", Toast.LENGTH_SHORT).show()
+            return
+        }
         val collection = firestore.collection("Users")
         val document = collection.document(userId)
         document.delete()
             .addOnSuccessListener {
                 Toast.makeText(MyApplication.getAppContext(), "User deleted successfully!", Toast.LENGTH_SHORT).show()
             }
-            .addOnFailureListener {
+            .addOnFailureListener { exception ->
                 Toast.makeText(MyApplication.getAppContext(), "User deletion failed!\nTry again", Toast.LENGTH_SHORT).show()
+                Log.e("DeleteUser", "Error deleting user with ID: $userId", exception)
             }
     }
 
     fun getUsersFromSharedPreferences() : LiveData<List<Users>> {
         val listOfUsers = mutableListOf<Users>()
-        val allEntries = sharedPreferences.all
-        for ((key, value) in allEntries) {
+        val allData = sharedPreferences.all
+        for ((key, value) in allData) {
             val userArray = (value as String).split(",")
             if (userArray.size == 6) {
                 val userId = userArray[0]
                 val userName = userArray[1]
                 val userImgProfile = userArray[2]
+                Toast.makeText(MyApplication.getAppContext(), "$userImgProfile", Toast.LENGTH_SHORT).show()
+                Log.d("hellothere", "$userImgProfile")
                 val userEmail = userArray[3]
-                val userPhone = userArray[4]
+                val userPhone = userArray[4].trim()
                 val userDob = userArray[5]
-                val user = Users(userId, userName, userImgProfile, userEmail, userPhone.toLong(), userDob)
-                listOfUsers.add(user)
+                try {
+                    val phone = userPhone.toLong()
+                    val user = Users(userId, userName, userImgProfile, userEmail, phone, userDob)
+                    listOfUsers.add(user)
+                } catch (e: NumberFormatException) {
+                    // Handle the NumberFormatException
+                }
             }
         }
         userListSharedPreferences.postValue(listOfUsers!!)
@@ -80,15 +93,17 @@ class ViewModel: ViewModel() {
     }
 
     fun getUserFromUserId(userId: String?, callback: (Users?) -> Unit) {
-        var user: Users? = null
-        firestore.collection("Users").document(userId!!).get()
-            .addOnSuccessListener { userDocument ->
-                if (userDocument!=null) {
-                    user = userDocument.toObject(Users::class.java)
-                    callback(user)
-                } else {
-                    callback(null)
+        if (userId != null) {
+            firestore.collection("Users").document(userId).get()
+                .addOnSuccessListener { userDocument ->
+                    if (userDocument!=null) {
+                        val user = userDocument.toObject(Users::class.java)
+                        Log.d("ImageCheck", "${user?.imgProfile}")
+                        callback(user)
+                    } else {
+                        callback(null)
+                    }
                 }
-            }
+        }
     }
 }
